feat(organisation): cleaned up some exceptions and renames test
