+++ src/core/domain/exceptions/models/Organisation/OrganisationName/OrganisationNameEmptyException.cs
@@ -1,5 +1,3 @@
-using System.Runtime.Serialization;
-
 namespace domain.exceptions.models.organisation;
 
 /// <summary>
